+++ src/main/java/life/qbic/subscriptions/ProjectLibrary.java
@@ -1,4 +1,4 @@
-package life.qbic.springbootresttemplate;
+package life.qbic.subscriptions;
 
 import java.util.ArrayList;
 import java.util.List;
