+++ src/main/java/life/qbic/subscriptions/ProjectRepository.java
@@ -1,4 +1,4 @@
-package life.qbic.springbootresttemplate;
+package life.qbic.subscriptions;
 
 import java.util.Optional;
