+++ src/main/java/life/qbic/subscriptions/encoding/DecodingException.java
@@ -1,4 +1,4 @@
-package life.qbic.springbootresttemplate.encoding;
+package life.qbic.subscriptions.encoding;
 
 public class DecodingException extends RuntimeException {
