+++ src/main/java/life/qbic/subscriptions/subscriptions/SubscriptionRepository.java
@@ -1,4 +1,4 @@
-package life.qbic.springbootresttemplate.subscriptions;
+package life.qbic.subscriptions.subscriptions;
 
 /**
  * <b>Interface SubscriptionRepository</b>
